Break out the `whoami ...` command hierarchy
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`whoami ...` hierarchy.
